requirements.txt
# Deep Learning and Computer Vision
torch>=1.11.0
torchvision>=0.12.0
opencv-python>=4.5.0
dlib>=19.24.0
face-recognition>=1.3.0
mtcnn>=0.1.1

# Web Framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Caching and Database
redis>=4.0.0
aioredis>=2.0.0

# Data Processing and Utilities
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=1.0.0

# Logging and Configuration
pydantic>=1.8.0
python-dotenv>=0.19.0
loguru>=0.5.3

# Development and Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# Additional utilities
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0